refactor: 로그인 한 사용자 정보 이름인 "login" 상수 관리
