+++ src/main/java/application/auth/Authority.java
@@ -1,12 +1,14 @@
-package application.filter.auth;
+package application.auth;
 
 import session.Session;
 
 public class Authority {
 
+    public static final String ATTRIBUTE_NAME_FOR_LOGIN = "login";
+
     public static void validateAuthority(Session session) throws UnauthorizedException {
         try {
-            if (session.getAttribute("login").equals(true)) {
+            if (session.getAttribute(ATTRIBUTE_NAME_FOR_LOGIN).equals(true)) {
                 return;
             }
             throw new UnauthorizedException();
